requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aom=3.5.0=h63175ca_0
blas=1.0=mkl
bzip2=1.0.8=he774522_0
ca-certificates=2023.01.10=haa95532_0
cairo=1.16.0=hd694305_1014
certifi=2022.12.7=py39haa95532_0
dav1d=1.0.0=h8ffe710_1
expat=2.5.0=h1537add_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.2=hbde0cde_0
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
freetype=2.12.1=ha860e81_0
gettext=0.21.1=h5728263_0
graphite2=1.3.14=hd77b12b_1
harfbuzz=6.0.0=he256f1b_0
icu=70.1=h0e60522_0
intel-openmp=2021.4.0=haa95532_3556
jpeg=9e=h2bbff1b_0
lerc=3.0=hd77b12b_0
libavif=0.11.1=h8f5c5bc_0
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_6
libflac=1.4.2=h63175ca_0
libglib=2.74.1=h79619a9_0
libiconv=1.17=h8ffe710_0
libmad=0.15.1b=hcfcfb64_1001
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h19919ed_0
libtiff=4.5.0=h6c2663c_1
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
libzlib=1.2.13=hcfcfb64_4
lz4-c=1.9.4=h2bbff1b_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
mpg123=1.31.2=h63175ca_0
numpy=1.23.5=py39h3b20f71_0
numpy-base=1.23.5=py39h4da318b_0
openssl=1.1.1t=h2bbff1b_0
pcre2=10.37=h0ff8eda_1
pillow=9.3.0=py39hd77b12b_2
pip=22.3.1=py39haa95532_0
pixman=0.40.0=h2bbff1b_1
portmidi=2.0.4=h63175ca_0
pygame=2.1.3=py39h928c5e4_0
python=3.9.16=h6244533_0
python_abi=3.9=2_cp39
sdl2=2.26.3=h63175ca_0
sdl2_image=2.6.3=h090e09f_0
sdl2_mixer=2.6.3=h090ff69_0
sdl2_ttf=2.20.2=h41be9cf_0
setuptools=65.6.3=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.1=h2bbff1b_0
tk=8.6.12=h2bbff1b_0
tzdata=2022g=h04d1e81_0
ucrt=10.0.20348.0=haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.34.31931=h4c5c07a_10
wheel=0.38.4=py39haa95532_0
wincertstore=0.2=py39haa95532_2
xz=5.2.10=h8cc25b3_1
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h19a0ad4_0
